fix(drivers): propagate five swallowed/mishandled error paths (#1131, #1132, #1133, #1134, #1135)
mhuv2_send() (mbox) now returns -EIO on the same ACCESS_READY timeout
mhuv2_set_enabled() already treats as failure, instead of ringing the
doorbell and reporting success. The legacy ipm_arm_mhuv2.c driver
(a separate defect, still consumed by the Secure Enclave services
path) gets its missing max_data_size_get op and rejects any transfer
size other than the hardware's fixed 4 bytes. The CC3501E TI SPI
transport now recovers from an arm_transfer() failure via the
existing g_resync_count-style reinit self-heal instead of stalling
READY low forever. gpu2d's software blend casts each packed uint8_t
component to uint32_t before shifting, removing a real (UBSan-caught,
reachable at alpha 0x80+) signed left-shift UB. rv3028c7_init() now
propagates a failed PORF-clear write instead of discarding it.

+++ src/backends/gpu2d/sw_fallback.c
@@ -195,11 +195,17 @@ static uint32_t _blend_px(uint32_t src, uint32_t dst, alp_gpu2d_blend_mode_t mod
 		uint32_t og  = (sg * sa + dg * ia + 127u) / 255u;
 		uint32_t ob  = (sb * sa + db * ia + 127u) / 255u;
 		uint32_t oa  = sa + (da * ia + 127u) / 255u;
-		return (_u8(oa) << 24) | (_u8(or_) << 16) | (_u8(og) << 8) | _u8(ob);
+		/* _u8() returns uint8_t; integer promotion makes the << operand a
+		 * signed int, and a high-bit byte (e.g. 0xFF) shifted into bit 24
+		 * is not representable as int -- signed-shift UB (UBSan-caught,
+		 * #1134). Cast to uint32_t first so the shift is unsigned. */
+		return ((uint32_t)_u8(oa) << 24) | ((uint32_t)_u8(or_) << 16) | ((uint32_t)_u8(og) << 8) |
+		       (uint32_t)_u8(ob);
 	}
 
 	case ALP_GPU2D_BLEND_ADDITIVE: {
-		return (_u8(sa + da) << 24) | (_u8(sr + dr) << 16) | (_u8(sg + dg) << 8) | _u8(sb + db);
+		return ((uint32_t)_u8(sa + da) << 24) | ((uint32_t)_u8(sr + dr) << 16) |
+		       ((uint32_t)_u8(sg + dg) << 8) | (uint32_t)_u8(sb + db);
 	}
 
 	case ALP_GPU2D_BLEND_MULTIPLY: {
@@ -207,7 +213,8 @@ static uint32_t _blend_px(uint32_t src, uint32_t dst, alp_gpu2d_blend_mode_t mod
 		uint32_t og  = (sg * dg + 127u) / 255u;
 		uint32_t ob  = (sb * db + 127u) / 255u;
 		uint32_t oa  = (sa * da + 127u) / 255u;
-		return (_u8(oa) << 24) | (_u8(or_) << 16) | (_u8(og) << 8) | _u8(ob);
+		return ((uint32_t)_u8(oa) << 24) | ((uint32_t)_u8(or_) << 16) | ((uint32_t)_u8(og) << 8) |
+		       (uint32_t)_u8(ob);
 	}
 
 	default:
